desktop: mpv: subs with matching audio
Please do not ask me why.
[Upstream docs](https://mpv.io/manual/stable/#options-subs-with-matching-audio) even claim this to be the default.
However if this is not set mpv will apparently play a file with japanese AV and multiple subtitles without actually selecting a subtitle track.
With this it properly respects the alang/slang settings and plays it with English subtitles.

As much as it pains me, thanks Reddit.…
https://www.reddit.com/r/mpv/comments/1ao0vbi/default_subtitle_language_is_not_working/

Signed-off-by: benaryorg <binary@benary.org>